[BugFix] fix mm rope (#7274)

+++ fastdeploy/worker/gpu_model_runner.py
@@ -1013,7 +1013,7 @@ class GPUModelRunner(ModelRunnerBase):
 
         if len(rope_3d_position_ids["position_ids_idx"]) > 0 and self.enable_mm:
             packed_position_ids = paddle.to_tensor(
-                np.concatenate(rope_3d_position_ids["position_ids_lst"]), dtype="int64"
+                np.concatenate(rope_3d_position_ids["position_ids_lst"]), dtype="float32"
             )
             rope_3d_lst = self.prepare_rope3d(
                 packed_position_ids,
